Create Random Password Generated List
I am trying to connect the randomly generated passwords to list1, in order to check that they have been used at a later date. Totally stuck on keeping a running list. The code puts the newly created code into the list, but doesn't keep the previous one from when it was run before.